WIP: estimateFee refactor
